更新 strhelper 由于 jsparser 的调整而修复 jsonparser 的编译

+++ sub-proj/JsonPP/src/strhelper.cpp
@@ -37,7 +37,7 @@ namespace sunjwbase
 		std::wstring wstrTo;
 		wchar_t *wszTo = new wchar_t[wstrLen + 1];
 		wszTo[wstrLen] = L'\0';
-		MultiByteToWideChar(codePage, 0, str.c_str(), -1, wszTo, wstrLen);
+		MultiByteToWideChar(codePage, 0, str.c_str(), -1, wszTo, (int)wstrLen);
 		wstrTo = wszTo;
 		delete[] wszTo;
 		return wstrTo;
